# openapi.yaml
openapi: 3.0.3
info:
  title: Event Organizer REST API
  description: REST API for event management, attendee registration, ticketing, check-in, and payments.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: localhost
    description: Relative URL for local development
security:
  - BearerAuth: []
paths:
...
  /api/orders:
    post:
      summary: Create Order / Ticket Purchase
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
                - quantity
                - paymentMethod
              properties:
                eventId:
                  type: integer
                  example: 1
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 6
                  example: 2
                paymentMethod:
                  type: string
                  example: BCA_VA
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Event purchased successfully
                  data:
                    type: object
        '400':
          description: ValidationError or BadRequestAppError (e.g. sold out)
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Unauthorized / Missing or Invalid Token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppErrorResponse'
...
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppErrorResponse'
...
    TooManyRequests:
      description: Too Many Requests / Rate Limit Exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppErrorResponse'
...
  schemas:
...
    AppErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Error message detail
            timestamp:
              type: string
              format: date-time
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
            text/csv:
              schema:
                type: string
                format: binary